added header files to compiler PATH

+++ tests/__init__.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <criterion/criterion.h>
 #include <criterion/hooks.h>
-#include "../src/inc/tokenizer_config.h"
+#include "tokenizer_config.h"
 
 token_config_t token_config;
 
@@ -12,9 +12,9 @@ ReportHook(PRE_ALL)(struct criterion_test_set *tests) {
 ReportHook(POST_ALL)(struct criterion_global_stats *stats) {
     int asserts_failed = stats->asserts_failed;
     printf("[\e[0;34m====\e[0m] \e[1mAsserts:\e[0m (passed: \e[0;32m%d\e[0m, failed: %s%d%s)\n",
-        stats->asserts_passed,
+        (int) stats->asserts_passed,
         asserts_failed == 0 ? "" : "\e[0;31m",
-        asserts_failed,
+        (int) asserts_failed,
         asserts_failed == 0 ? "" : "\e[0m"
     );
 }
